Refactor maxNodeId in distset to uint64
It now matches the type of the nodeId.

+++ shard/index/dispatch.go
@@ -105,7 +105,7 @@ func Dispatch(
 	cm *cache.Manager,
 	cacheRoot string,
 	indexSchema models.IndexSchema,
-	maxNodeId uint,
+	maxNodeId uint64,
 	changes <-chan IndexPointChange,
 ) error {
 	// ---------------------------
