+++ shard/shard.go
@@ -209,7 +209,7 @@ func (s *Shard) InsertPoints(points []models.Point) error {
 		wg.Add(1)
 		go func() {
 			defer wg.Done()
-			err := index.Dispatch(ctx, bm, s.cacheManager, s.dbFile, s.collection.IndexSchema, uint(s.maxNodeId.Load()), indexQ)
+			err := index.Dispatch(ctx, bm, s.cacheManager, s.dbFile, s.collection.IndexSchema, s.maxNodeId.Load(), indexQ)
 			if err != nil {
 				cancel(fmt.Errorf("could not dispatch index: %w", err))
 			}
@@ -292,7 +292,7 @@ func (s *Shard) UpdatePoints(points []models.Point) ([]uuid.UUID, error) {
 		wg.Add(1)
 		go func() {
 			defer wg.Done()
-			err := index.Dispatch(ctx, bm, s.cacheManager, s.dbFile, s.collection.IndexSchema, uint(s.maxNodeId.Load()), indexQ)
+			err := index.Dispatch(ctx, bm, s.cacheManager, s.dbFile, s.collection.IndexSchema, s.maxNodeId.Load(), indexQ)
 			if err != nil {
 				cancel(fmt.Errorf("could not dispatch index: %w", err))
 			}
@@ -379,7 +379,7 @@ func (s *Shard) SearchPoints(searchRequest models.SearchRequest) ([]models.Searc
 			return fmt.Errorf("could not get points bucket: %w", err)
 		}
 		// ---------------------------
-		rSet, results, err := index.Search(context.Background(), bm, s.cacheManager, s.dbFile, s.collection.IndexSchema, uint(s.maxNodeId.Load()), searchRequest.Query)
+		rSet, results, err := index.Search(context.Background(), bm, s.cacheManager, s.dbFile, s.collection.IndexSchema, s.maxNodeId.Load(), searchRequest.Query)
 		if err != nil {
 			return fmt.Errorf("could not perform search: %w", err)
 		}
@@ -514,7 +514,7 @@ func (s *Shard) DeletePoints(deleteSet map[uuid.UUID]struct{}) ([]uuid.UUID, err
 		wg.Add(1)
 		go func() {
 			defer wg.Done()
-			err := index.Dispatch(ctx, bm, s.cacheManager, s.dbFile, s.collection.IndexSchema, uint(s.maxNodeId.Load()), indexQ)
+			err := index.Dispatch(ctx, bm, s.cacheManager, s.dbFile, s.collection.IndexSchema, s.maxNodeId.Load(), indexQ)
 			if err != nil {
 				cancel(fmt.Errorf("could not dispatch index: %w", err))
 			}
